Rename submodules under `cupyx.scipy.linalg` package
