+++ cupyx/scipy/linalg/__init__.py
@@ -1,3 +1,3 @@
-from cupyx.scipy.linalg.special_matrices import *  # NOQA
-from cupyx.scipy.linalg.solve_triangular import solve_triangular  # NOQA
-from cupyx.scipy.linalg.decomp_lu import lu_factor, lu_solve  # NOQA
+from cupyx.scipy.linalg._special_matrices import *  # NOQA
+from cupyx.scipy.linalg._solve_triangular import solve_triangular  # NOQA
+from cupyx.scipy.linalg._decomp_lu import lu_factor, lu_solve  # NOQA
