+++ cupyx/scipy/linalg/_special_matrices.py
